Refactoring DockerInformation
